requirements.txt
pandas==3.0.2
numpy==2.4.4
python-dotenv==1.2.2
google-api-python-client==2.196.0
google-auth-httplib2==0.4.0
google-auth-oauthlib==1.4.0
gspread==6.2.1
pyarrow==24.0.0
openpyxl==3.1.5
gigachat==0.2.1
